Upload template
Starting logic for a monorepo format that allows independent development of nextjs applications that are then integrated into a core platform as first-class-citizens. For sanity in developing new features for large projects, and multi-developer collaboration on different features, without losing interoperability. Able to test independent and integrated operation in a development environment.